requirements.txt
asgiref==3.8.1
django-jquery==3.1.0
django-model-utils==4.5.1
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1